Move thread service into its own module
