+++ app/api/api.go
@@ -11,6 +11,7 @@ import (
 	"github.com/France-ioi/AlgoreaBackend/app/api/currentuser"
 	"github.com/France-ioi/AlgoreaBackend/app/api/groups"
 	"github.com/France-ioi/AlgoreaBackend/app/api/items"
+	"github.com/France-ioi/AlgoreaBackend/app/api/threads"
 	"github.com/France-ioi/AlgoreaBackend/app/api/users"
 	"github.com/France-ioi/AlgoreaBackend/app/database"
 	"github.com/France-ioi/AlgoreaBackend/app/domain"
@@ -40,6 +41,7 @@ func Router(db *database.DB, serverConfig, authConfig *viper.Viper, domainConfig
 	r.Group((&auth.Service{Base: srv}).SetRoutes)
 	r.Group((&contests.Service{Base: srv}).SetRoutes)
 	r.Group((&items.Service{Base: srv}).SetRoutes)
+	r.Group((&threads.Service{Base: srv}).SetRoutes)
 	r.Group((&groups.Service{Base: srv}).SetRoutes)
 	r.Group((&answers.Service{Base: srv}).SetRoutes)
 	r.Group((&currentuser.Service{Base: srv}).SetRoutes)
