+++ app/api/threads/bdd_test.go
@@ -1,0 +1,13 @@
+// +build !unit
+
+package threads_test
+
+import (
+	"testing"
+
+	"github.com/France-ioi/AlgoreaBackend/testhelpers"
+)
+
+func TestBDD(t *testing.T) {
+	testhelpers.RunGodogTests(t, "")
+}
